Diabetes
The dataset is part of a large dataset held at the National Institutes of Diabetes-Digestive-Kidney Diseases in the US.
Data used for a diabetes survey of Pima Indian women aged 21 years and older living in Phoenix, the 5th largest city in the State of Arizona. The target variable is “outcome”, where 1 indicates a positive diabetes test result and 0 indicates a negative test result.